+++ packages/docusaurus-babel/package.json
@@ -36,8 +36,10 @@
     "@babel/preset-typescript": "^7.22.5",
     "@babel/runtime": "^7.22.6",
     "@babel/runtime-corejs3": "^7.22.6",
+    "@docusaurus/utils": "3.5.2",
     "babel-loader": "^9.1.3",
-    "babel-plugin-dynamic-import-node": "^2.3.3"
+    "babel-plugin-dynamic-import-node": "^2.3.3",
+    "fs-extra": "^11.1.1"
   },
   "engines": {
     "node": ">=18.0"
