+++ packages/docusaurus-babel/src/index.ts
@@ -5,4 +5,4 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-export const x = 42;
+export {getCustomBabelConfigFilePath, getBabelOptions} from './utils';
