+++ packages/docusaurus-bundler/package.json
@@ -33,7 +33,6 @@
     "postcss": "^8.4.26",
     "postcss-loader": "^7.3.3",
     "file-loader": "^6.2.0",
-    "fs-extra": "^11.1.1",
     "mini-css-extract-plugin": "^2.9.1",
     "null-loader": "^4.0.1",
     "react-dev-utils": "^12.0.1",
