+++ packages/docusaurus-bundler/src/loaders/jsLoader.ts
@@ -5,48 +5,9 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import fs from 'fs-extra';
-import path from 'path';
-import {BABEL_CONFIG_FILE_NAME} from '@docusaurus/utils';
+import {getBabelOptions} from '@docusaurus/babel';
 import {importSwcJsLoaderFactory} from '../importFaster';
 import type {ConfigureWebpackUtils, DocusaurusConfig} from '@docusaurus/types';
-import type {TransformOptions} from '@babel/core';
-
-export async function getCustomBabelConfigFilePath(
-  siteDir: string,
-): Promise<string | undefined> {
-  const customBabelConfigurationPath = path.join(
-    siteDir,
-    BABEL_CONFIG_FILE_NAME,
-  );
-  return (await fs.pathExists(customBabelConfigurationPath))
-    ? customBabelConfigurationPath
-    : undefined;
-}
-
-export function getBabelOptions({
-  isServer,
-  babelOptions,
-}: {
-  isServer?: boolean;
-  babelOptions?: TransformOptions | string;
-} = {}): TransformOptions {
-  if (typeof babelOptions === 'string') {
-    return {
-      babelrc: false,
-      configFile: babelOptions,
-      caller: {name: isServer ? 'server' : 'client'},
-    };
-  }
-  return {
-    ...(babelOptions ?? {
-      presets: [require.resolve('@docusaurus/babel/preset')],
-    }),
-    babelrc: false,
-    configFile: false,
-    caller: {name: isServer ? 'server' : 'client'},
-  };
-}
 
 const BabelJsLoaderFactory: ConfigureWebpackUtils['getJSLoader'] = ({
   isServer,
