+++ packages/docusaurus-mdx-loader/README.md
@@ -18,7 +18,6 @@ module: {
     {
       test: /\.mdx?$/,
       use: [
-        'babel-loader',
         {
           loader: '@docusaurus/mdx-loader',
           options: {
