+++ packages/docusaurus/package.json
@@ -36,6 +36,7 @@
     "@babel/core": "^7.23.3",
     "@babel/generator": "^7.23.3",
     "@babel/traverse": "^7.22.8",
+    "@docusaurus/babel": "3.5.2",
     "@docusaurus/bundler": "3.5.2",
     "@docusaurus/logger": "3.5.2",
     "@docusaurus/mdx-loader": "3.5.2",
